ci: build pull requests in copr
New copr project will be created for each pull request under the sssd
group. The project will be automatically deleted after 60 days or after
the pull request is closed.

The workflow also produces comment on the pull request, there may
be a little race condition between the time where the comment is
already visible and the time when the project is created, but it is
not long. The reason for this is that the comment is created before
the project is built in order for it to be always on the top.

The workflow works like this:
- submit comment to the pull request
- build tarball and spec file (rvn is sssd-pr#prid-#runid)
- build srpm
- upload srpm as an artifact to the job
- create copr project @sssd/pr#prid if not exist
- cancel previous pending/running builds if there are any
- build package in centos-* and fedora-*

The workflow fails if build in any chroot fails.
The workflow succeeds if all chroots are successful.